Rename ejercicio 13 to Ejercicio 13
Exercise 13: Write a program that chooses a random number between 1
and 100. The user then tries to guess that number. The program will tell
the user whether the hidden number is higher or lower than the number that
was guessed. When the correct number is guessed, report back how many
guesses it took to find the correct number. Then offer to play the game
again